extract reusable types
